+++ spec/rack_spec.rb
@@ -1,0 +1,6 @@
+require "rack/test"
+require_relative '../lib/racker'
+
+
+
+end
